Big refactor
Move things around for simpler reading later
